Sloppy commit message:
Thought I had committed another version where src was set to (dst =
arraylen - 1) - 1 where it would trail back, thinking that skipping over
numbers/bits already "in place" would be faster (pretty sure it is,
still).

Noticed that the algorithm didnt actually work only after installing
linux on a VM to use perf to try to optimise and compare with commented
assembly - RAND_MAX is set to just 2^15 on my windows machine, and in an
attempt to simulate higher random numbers I just bit shifted them left
by 10, which played well by the "trailing algorithm".

Never ended up making it work properly, and had issues with not
repeating the "skipping" loop before and inside the main while loop
(while src<dst).

From what I could remember, the first commit performance was about on
par with the broken one, but also after considering a sequence akin to
"0 1 1 1 1 1 1 0 1" where the first "one" left of the "zero" closeset to
the end of the array would be shifted repeatedly (wastefully) back - so
instead it made sense for the source element to be initialised to 0 and
increment up.

Also dropped the check for valid pointer after checking if qsort was a
void function. While I would like to rewrite/omit the arraylen > 1
condition in the main while loop, the performance decreased for every
attempt.
Same thing happened with the prior algorithm.

Using the _Generic keyword/macros to make the functions available to any
integral type would be interesting, though would have to give thought to
sorting negative integers. Renamed the identifier macro in the
highestpower header to acknowledge that the highest bit wouldnt always
translate to "greatest" bit (ie, positive bit > negative bit/s).

The new algorithm initialises the start and endpoints, decrements the
endpoint until it finds an index that holds a value not sharing the same
bit, increments the startpoint to the opposite effect (until beyond the
last vacant index), where if there are vacant spaces on the right side
of the array that can be occupied by elements with a present bit on the
left, the will be swapped. The would also be no vacant elements on the
right side if the source element == destination element.

Perf report seemed normal with the comparisons taking a shared majority
of the runtime alongside what I think was pushing the incrementors and
the temporary variable.

In the swap function, goto acts as a gate / a replacement for while
loops not having complex expressions (or me being clueless). Didn't want
to write a procedure, though I don't think that would've worked either.

Maybe I should've gone to college. Or something